# src/main/resources/api.yml
openapi: 3.0.3
info:
  title: Products Microservice API
  description: Microservice for product catalog
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Development server

tags:
  - name: products
    description: Product operations

paths:
  /products:
    get:
      tags:
        - products
      summary: Get all products
      responses:
        '200':
          description: Lista de productos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
              examples:
                ejemplo:
                  value:
                    - id: "64f25d1c8b9f1a7c2d123456"
                      name: "Laptop Dell XPS"
                      price: 1200.50
                      stock: 10
                    - id: "64f25d1c8b9f1a7c2d123457"
                      name: "iPhone 14"
                      price: 999.99
                      stock: 5
    post:
      tags:
        - products
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
            example:
              name: "Laptop Dell XPS"
              price: 1200.50
              stock: 10
      responses:
        '201':
          description: Producto creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'

  /products/{id}:
    get:
      tags:
        - products
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found

    put:
      tags:
        - products
      summary: Update product (except stock)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdateRequest'
            example:
              name: "Laptop Dell XPS 13"
              price: 1100.00
      responses:
        '200':
          description: Producto actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found

    delete:
      tags:
        - products
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Producto eliminado exitosamente
        '404':
          description: Product not found

  /products/{id}/addStock:
    patch:
      tags:
        - products
      summary: Add stock to product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: quantity
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Stock added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found

  /products/{id}/purchase:
    post:
      tags:
        - products
      summary: Purchase product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: quantity
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Compra realizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Error - No hay suficiente stock
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: "Not enough stock for product with id 64f25d1c8b9f1a7c2d123456"
        '404':
          description: Product not found

components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: number
          format: double
        stock:
          type: integer

    ProductRequest:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
          format: double
        stock:
          type: integer
      required: [name, price, stock]

    ProductUpdateRequest:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
      required: [name, price]